refactor file names to kebab-case
