+++ src/app/dashboard/(overview)/page.tsx
@@ -1,4 +1,4 @@
-import DashboardPage from "@/components/pages/dashboard/DashboardPage"
+import DashboardPage from "@/components/pages/dashboard/dashboard-page"
 
 export default function Overview() {
   return (
