+++ src/app/dashboard/settings/loading.tsx
@@ -1,5 +1,5 @@
-import DashboardPage from "@/components/pages/dashboard/DashboardPage"
-import { LoadingSettingsForm } from "@/components/pages/dashboard/settings/SettingsForm"
+import DashboardPage from "@/components/pages/dashboard/dashboard-page"
+import { LoadingSettingsForm } from "@/components/pages/dashboard/settings/settings-form"
 
 export default function LoadingSettings() {
   return (
