+++ src/app/dashboard/settings/page.tsx
@@ -2,10 +2,10 @@ import { Suspense } from "react"
 
 import { getSession } from "@/lib/auth/get-session"
 import { getUser } from "@/lib/query/user.queries"
-import DashboardPage from "@/components/pages/dashboard/DashboardPage"
+import DashboardPage from "@/components/pages/dashboard/dashboard-page"
 import SettingsForm, {
   LoadingSettingsForm,
-} from "@/components/pages/dashboard/settings/SettingsForm"
+} from "@/components/pages/dashboard/settings/settings-form"
 
 export default async function SettingsPage() {
   const session = await getSession()
